Часть логики и представления из MessageScreen вынесены в компонент MessageInput. При формировании сообщения имя автора берётся из объекта чата.
