+++ lesson5/src/store/actionCreators/RemoveChatActionCreator.js
@@ -1,0 +1,8 @@
+import { CHATS_REMOVE_CHAT } from "../chats/action";
+
+export const RemoveChatActionCreator = (chatID) => {
+  return {
+    type: CHATS_REMOVE_CHAT,
+    id: chatID,
+  };
+};
