+++ lesson5/src/store/chats/action.js
@@ -1,2 +1,3 @@
 export const CHATS_ADD_CHAT = "CHATS_ADD_CHAT";
+export const CHATS_REMOVE_CHAT = "CHATS_REMOVE_CHAT";
 export const GET_CHATS_LIST = "GET_CHATS_LIST";
