+++ lesson5/src/store/chats/reducer.js
@@ -1,5 +1,5 @@
 import { nanoid } from "nanoid";
-import { CHATS_ADD_CHAT, GET_CHATS_LIST } from "./action";
+import { CHATS_ADD_CHAT, CHATS_REMOVE_CHAT, GET_CHATS_LIST } from "./action";
 import { chats } from "../../imit_chats/imit_chats";
 
 const initialChatList = {
@@ -12,6 +12,16 @@ export const chatsReducer = (state = initialChatList, action) => {
       return {
         ...state,
         chatList: [...state.chatList, { name: action.name, id: nanoid() }],
+      };
+    }
+    case CHATS_REMOVE_CHAT: {
+      return {
+        ...state,
+        chatList: [
+          state.chatList.filter((chat) => {
+            return chat.id !== action.id;
+          }),
+        ],
       };
     }
     case GET_CHATS_LIST: {
